src/schooltool/generations/evolve10.py
#
# SchoolTool - common information systems platform for school administration
# Copyright (c) 2006 Shuttleworth Foundation
#
# This program is free software; you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation; either version 2 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program; if not, write to the Free Software
# Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
#
"""
Upgrade SchoolTool to generation 10.

Put a timezone attribute on all timetables and school timetables.
Take the timezone out of the schoolwide setting.

$Id$
"""

from zope.app.publication.zopepublication import ZopePublication
from zope.app.generations.utility import findObjectsProviding

from schooltool.app.interfaces import ISchoolToolApplication
from schooltool.app.interfaces import IApplicationPreferences
from schooltool.timetable.interfaces import ITimetables, IHaveTimetables
from schooltool.group.group import Group


def evolve(context):
    root = context.connection.root()[ZopePublication.root_name]
    for app in findObjectsProviding(root, ISchoolToolApplication):
        tz = IApplicationPreferences(app).timezone

        for tts in app['ttschemas'].values():
            tts.timezone = tz

        for tt in ITimetables(app).timetables.values():
            tt.timezone = tz

        for obj in findObjectsProviding(root, IHaveTimetables):
            for tt in ITimetables(obj).timetables.values():
                tt.timezone = tz
